update the code type into int from byte, and other tiny improvements

+++ computer-test/src/main/java/com/baidu/hugegraph/computer/core/common/ValueTypeTest.java
@@ -60,7 +60,7 @@ public class ValueTypeTest {
     @Test
     public void testException() {
         Assert.assertThrows(ComputerException.class, () -> {
-            ValueType.fromCode((byte) -100);
+            ValueType.fromCode(-100);
         });
     }
 }
